+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.5.0] - 2021-04-15
+
 ### Changed
 
-- Latest SCPG schema applied with deprecated properties removed
+- Latest SCPG schema applied with deprecated properties removed.
+- Removed deprecated `DeltaGraph.apply`
 
 ## [0.4.7] - 2021-04-15